adds a newcommand and newevent method
